# services/api/openapi.json
{
  "components": {
    "schemas": {
      "ActionType": {
        "enum": [
          "investigate_further",
          "contain_host_manual",
          "reset_credentials_manual",
          "block_ioc_manual",
          "tune_rule",
          "close_as_fp",
          "notify_user"
        ],
        "title": "ActionType",
        "type": "string"
      },
...
      "CaseReport": {
        "additionalProperties": false,
        "description": "Case output of the Reporting agent. Turkish fields end in `_tr`.\n\n`data_gaps` comes from `AgentResult`.",
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "title": "Claims",
            "type": "array"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "data_gaps": {
            "items": {
              "$ref": "#/components/schemas/DataGap"
            },
            "title": "Data Gaps",
            "type": "array"
          },
          "injection_suspected": {
            "title": "Injection Suspected",
            "type": "boolean"
          },
          "notify_level": {
            "$ref": "#/components/schemas/Level"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "maxItems": 8,
            "title": "Recommendations",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "summary_tr": {
            "maxLength": 600,
            "title": "Summary Tr",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "urgent_events": {
            "items": {
              "$ref": "#/components/schemas/UrgentEvent"
            },
            "maxItems": 15,
            "title": "Urgent Events",
            "type": "array"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "verdict": {
            "$ref": "#/components/schemas/CaseVerdict"
          }
        },
        "required": [
          "task_id",
          "status",
          "claims",
          "data_gaps",
          "injection_suspected",
          "usage",
          "summary_tr",
          "verdict",
          "confidence",
          "notify_level",
          "urgent_events",
          "recommendations"
        ],
        "title": "CaseReport",
        "type": "object"
      },
...
      "CaseStatus": {
        "description": "`cases.status`.",
        "enum": [
          "running",
          "decided",
          "no_ai_decision",
          "closed"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
...
      "CaseVerdict": {
        "enum": [
          "tp",
          "fp",
          "suspicious"
        ],
        "title": "CaseVerdict",
        "type": "string"
      },
...
      "Claim": {
        "additionalProperties": false,
        "properties": {
          "evidence_ids": {
            "items": {
              "pattern": "^ev_\\S+$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "text": {
            "maxLength": 300,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "evidence_ids"
        ],
        "title": "Claim",
        "type": "object"
      },
      "Confidence": {
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Confidence",
        "type": "string"
      },
...
      "DataGap": {
        "additionalProperties": false,
        "properties": {
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/DataGapReason"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "period_start",
          "period_end",
          "reason"
        ],
        "title": "DataGap",
        "type": "object"
      },
      "DataGapReason": {
        "enum": [
          "no_data",
          "not_parsed",
          "not_visible",
          "query_failed",
          "budget_exhausted"
        ],
        "title": "DataGapReason",
        "type": "string"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "GroupDetail": {
        "additionalProperties": false,
        "description": "`GET /groups/{group_id}`: the group, its offenses and its evaluation's decision.",
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "case_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "group": {
            "$ref": "#/components/schemas/GroupSummary"
          },
          "notify_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Level"
              },
              {
                "type": "null"
              }
            ]
          },
          "offenses": {
            "items": {
              "$ref": "#/components/schemas/GroupOffense"
            },
            "title": "Offenses",
            "type": "array"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseVerdict"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "group"
        ],
        "title": "GroupDetail",
        "type": "object"
      },
      "GroupOffense": {
        "additionalProperties": false,
        "description": "One offense of a group.",
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "offense_id": {
            "title": "Offense Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/OffenseStatus"
          }
        },
        "required": [
          "offense_id",
          "description",
          "status",
          "first_seen_at"
        ],
        "title": "GroupOffense",
        "type": "object"
      },
      "GroupStatus": {
        "description": "`offense_groups.status`.",
        "enum": [
          "open",
          "storm",
          "closed"
        ],
        "title": "GroupStatus",
        "type": "string"
      },
      "GroupSummary": {
        "additionalProperties": false,
        "description": "One row of the group list.",
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "group_id": {
            "title": "Group Id",
            "type": "string"
          },
          "offense_count": {
            "title": "Offense Count",
            "type": "integer"
          },
          "rule_set_hash": {
            "title": "Rule Set Hash",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/GroupStatus"
          },
          "window_end": {
            "format": "date-time",
            "title": "Window End",
            "type": "string"
          },
          "window_start": {
            "format": "date-time",
            "title": "Window Start",
            "type": "string"
          }
        },
        "required": [
          "group_id",
          "rule_set_hash",
          "window_start",
          "window_end",
          "offense_count",
          "status"
        ],
        "title": "GroupSummary",
        "type": "object"
      },
...
      "Level": {
        "description": "Severity level. Values compare as strings, not by severity.",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Level",
        "type": "string"
      },
      "Me": {
        "additionalProperties": false,
        "description": "`GET /me`: who the session is and what it may do.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "title": "Roles",
            "type": "array"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "display_name",
          "roles"
        ],
        "title": "Me",
        "type": "object"
      },
...
      "OffenseStatus": {
        "description": "`offenses_seen.status`.",
        "enum": [
          "pending",
          "running",
          "done",
          "skipped",
          "grouped"
        ],
        "title": "OffenseStatus",
        "type": "string"
      },
...
      "Page_GroupSummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GroupSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[GroupSummary]",
        "type": "object"
      },
...
      "PlatformFlag": {
        "description": "`platform_flags.name`: the switches the platform knows (T-23).",
        "enum": [
          "writes_enabled"
        ],
        "title": "PlatformFlag",
        "type": "string"
      },
      "PlatformFlagState": {
        "additionalProperties": false,
        "description": "`GET /admin/platform-flags`: every known flag. A flag with no row is off.",
        "properties": {
          "changed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed At"
          },
          "changed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "$ref": "#/components/schemas/PlatformFlag"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "name",
          "enabled"
        ],
        "title": "PlatformFlagState",
        "type": "object"
      },
      "PlatformFlagUpdate": {
        "additionalProperties": false,
        "description": "`PUT /admin/platform-flags/{name}` (admin). `reason` is required and may not be blank.",
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 500,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "reason"
        ],
        "title": "PlatformFlagUpdate",
        "type": "object"
      },
...
      "Problem": {
        "description": "The shape of an RFC 9457 answer, for the UI's own error handling.\n\nNot a response model of any route: `ais0c_api.openapi` makes it every operation's `default`\nresponse in the schema. `errors`, `domains` and `list_names` appear only on the problems that\ncarry them.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Domains"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "list_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "List Names"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "Recommendation": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "evidence_ids": {
            "items": {
              "pattern": "^ev_\\S+$",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "rationale": {
            "maxLength": 300,
            "title": "Rationale",
            "type": "string"
          },
          "target": {
            "maxLength": 200,
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "action_type",
          "target",
          "rationale",
          "evidence_ids"
        ],
        "title": "Recommendation",
        "type": "object"
      },
      "Role": {
        "description": "`api.md`'s three roles, least to most.",
        "enum": [
          "operator",
          "hunter",
          "admin"
        ],
        "title": "Role",
        "type": "string"
      },
      "RunStatus": {
        "enum": [
          "completed",
          "budget_exhausted",
          "failed"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "Usage": {
        "additionalProperties": false,
        "properties": {
          "seconds": {
            "title": "Seconds",
            "type": "number"
          },
          "tokens": {
            "title": "Tokens",
            "type": "integer"
          },
          "tool_calls": {
            "title": "Tool Calls",
            "type": "integer"
          }
        },
        "required": [
          "tokens",
          "tool_calls",
          "seconds"
        ],
        "title": "Usage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The analyst UI's API. The UI talks only to this service.\n\nErrors are RFC 9457 problems whose `title` is a machine-readable code (`catalog.rule_not_found`,\n`qa.already_resolved`); the Turkish message the user reads is built by the UI from that code.\nTimes are ISO 8601 in UTC. Lists are cursor-paged with `?cursor=&limit=`.\n",
    "title": "ais0c analyst API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/platform-flags": {
      "get": {
        "description": "Every flag the platform knows, on or off. A flag with no row is off and `changed_by` empty.\n\nToday the only flag is the kill switch `writes_enabled` (T-23).",
        "operationId": "get_flags_api_v1_admin_platform_flags_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlatformFlagState"
                  },
                  "title": "Response Get Flags Api V1 Admin Platform Flags Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Flags",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/v1/admin/platform-flags/{name}": {
      "put": {
        "description": "Switch a platform flag (admin). `reason` is required and may not be blank. An unknown flag\nname is a 404 and nothing is written.",
        "operationId": "put_flag_api_v1_admin_platform_flags__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformFlagUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformFlagState"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Put Flag",
        "tags": [
          "monitoring"
        ]
      }
    },
...
    "/api/v1/groups": {
      "get": {
        "description": "The groups, newest window first. `from` and `to` bound `window_start`.",
        "operationId": "get_groups_api_v1_groups_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/GroupStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_GroupSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Groups",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{group_id}": {
      "get": {
        "description": "The group, the offenses it holds and the decision of its own case.\n\nA group with no case yet (still being filled) answers with the group and its offenses; an\nunknown `group_id` is a 404.",
        "operationId": "get_group_api_v1_groups__group_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Group",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "description": "The session's subject, display name and roles, most privileged first.",
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Me",
        "tags": [
          "monitoring"
        ]
      }
    },
...
  }
}